Support for JMP instructions + bug fix
Supports absolute and indirect JMP instructions. Also (I think) fixed a
potential bug in a few places by forcing the ALU carry input to 0
whenever we're just adding zero to an address to store it - otherwise,
if the status reg carry bit was set, we'd end up adding 1 instead of
just storing the byte.