Support for JSR op
Now supports jumping to subroutines (though not returning from them as
yet). Had to make a change to the stack pointer register to make it work
- according to visual6502.org, S is actually a 2-stage register,
latching data in on ph1 and making it available as output on ph2 (so you
can read and write on the same cycle without overwriting the previous
value in the middle).

There's also some (probably unnecessary) complications in holding the
address low-byte in the CPU for 5 cycles. Following an idea from
6502.org, ADL is stored in the S register, while the stack pointer is
decremented in the ALU (unlike in other ops, where it's dec/incremented
internal to the S register). I'm not sure how the actual chip does it,
but this seems to work.