Testing done
Interrupt timing is tested, and seems to be working. Had to fix a few
bugs: (1) interrupts should be disabled by default (this will eventually
be done in the restart routine, but that's not implemented yet), (2) the
instruction control should only set IR to BRK on cycle 1 (previously, it
could happen anytime). Next up: implementing RTI