Support for ROR op
All addressing modes. This was a bit cheaty - I added a separate ROR
instruction to the ALU, since I wasn't sure exactly how the original did
it. It seems to work though, but I might change it later. Also, changed
the format of the AC wave in the simulator to hex since I was tired of
doing it manually.