Testing and debugging
Ok, so I've actually implemented a way to get a program to the CPU for
testing and debugging. The top file, NES, contains an instantiated CPU,
along with a UART and FIFO buffer. It communicates with a program on a
host PC to send/receive data and send status updates (containing
register info). The host PC controls all program execution